Fix parse code
